News: Add bitbox to news
